requirements.txt
asn1crypto==0.22.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==1.9
ddt==1.1.1
enum34==1.1.6
Flask==1.0.2
gevent==1.2.2
greenlet==0.4.13
idna==2.6
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.10
keyring==10.4.0
keyrings.alt==2.2
locustio==0.8.1
MarkupSafe==1.0
msgpack-python==0.5.6
pkg-resources==0.0.0
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.6.1
pycryptodomex==3.6.1
PyJWT==1.6.1
pyxdg==0.25
pyzmq==17.0.0
requests==2.18.4
SecretStorage==2.3.1
selenium==3.4.3
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.14.1

